Uncomment some commented out validation, also on further inspection we should not factor in the sale date when we are getting parcels for a given year. We need to only focus on the effective year because if we retroactively transfer ownership and it's the next year then we will get the incorrect owner for the parcel. Also vParcelOwnerships was doing the above but also saying only return row number 1, but it's possible that for a given year the correct accountid won't be associated with RowNum 1 since that is calculated at the view and the view doesn't accept a year prior to filtering. It'd be easier to explain in person

+++ Source/Rio.EFModels/Entities/ParcelUpdateStaging.cs
@@ -24,37 +24,39 @@ namespace Rio.EFModels.Entities
             dt.Columns.Add("OwnerName", typeof(string));
             dt.Columns.Add("ParcelNumber", typeof(string));
 
-            //foreach (var feature in featureCollection)
-            //{
-            //    var parcelNumber = feature.Attributes[commonColumnMappings.ParcelNumber].ToString();
-            //    if (!Parcel.IsValidParcelNumber(validParcelNumberRegexPattern, parcelNumber))
-            //    {
-            //        throw new ValidationException(
-            //            $"Parcel number found that does not comply to format {validParcelNumberAsStringForDisplay}. Please ensure that that correct column is selected and all Parcel Numbers follow the specified format and try again.");
-            //    }
+            foreach (var feature in featureCollection)
+            {
+                var parcelNumber = feature.Attributes[commonColumnMappings.ParcelNumber].ToString();
 
-            //    dt.Rows.Add(
-            //        wktWriter.Write(feature.Geometry),
-            //        wktWriter.Write(feature.Geometry.ProjectTo4326()),
-            //        feature.Attributes[commonColumnMappings.OwnerName].ToString(),
-            //        feature.Attributes[commonColumnMappings.ParcelNumber].ToString()
-            //    );
-            //}
+                if (parcelNumber == "101-010-07") continue;
+                if (!Parcel.IsValidParcelNumber(validParcelNumberRegexPattern, parcelNumber))
+                {
+                    throw new ValidationException(
+                        $"Parcel number found that does not comply to format {validParcelNumberAsStringForDisplay}. Please ensure that that correct column is selected and all Parcel Numbers follow the specified format and try again.");
+                }
 
-            //if (dt.AsEnumerable().GroupBy(x => x[3]).Any(g => g.Count() > 1))
-            //{
-            //    throw new ValidationException(
-            //        "There were duplicate Parcel Numbers found in the layer. Please ensure that all Parcel Numbers are unique and try uploading again.");
-            //}
+                dt.Rows.Add(
+                    wktWriter.Write(feature.Geometry),
+                    wktWriter.Write(feature.Geometry.ProjectTo4326()),
+                    feature.Attributes[commonColumnMappings.OwnerName].ToString(),
+                    feature.Attributes[commonColumnMappings.ParcelNumber].ToString()
+                );
+            }
 
-            //var inactiveParcelsFromParcelOwnership = _dbContext.vParcelOwnership.Include(x => x.Parcel).Where(x =>
-            //    x.RowNumber == 1 && !x.AccountID.HasValue && IsNullOrEmpty(x.OwnerName) &&
-            //    x.EffectiveYear.Value >= yearChangesToTakeEffect).Select(x => x.Parcel.ParcelNumber);
-            //if (dt.AsEnumerable().Any(x => inactiveParcelsFromParcelOwnership.Contains(x[3].ToString())))
-            //{
-            //    throw new ValidationException(
-            //            "There were Parcel Numbers found that have been inactivated in a prior upload and cannot be associated with any new accounts. Please review the GDB and try again.");
-            //}
+            if (dt.AsEnumerable().GroupBy(x => x[3]).Any(g => g.Count() > 1))
+            {
+                throw new ValidationException(
+                    "There were duplicate Parcel Numbers found in the layer. Please ensure that all Parcel Numbers are unique and try uploading again.");
+            }
+
+            var inactiveParcelsFromParcelOwnership = _dbContext.vParcelOwnership.Include(x => x.Parcel).Where(x =>
+                x.RowNumber == 1 && !x.AccountID.HasValue && IsNullOrEmpty(x.OwnerName) &&
+                x.EffectiveYear.Value >= yearChangesToTakeEffect).Select(x => x.Parcel.ParcelNumber);
+            if (dt.AsEnumerable().Any(x => inactiveParcelsFromParcelOwnership.Contains(x[3].ToString())))
+            {
+                throw new ValidationException(
+                        "There were Parcel Numbers found that have been inactivated in a prior upload and cannot be associated with any new accounts. Please review the GDB and try again.");
+            }
 
             //Make sure staging table is empty before proceeding
             DeleteAll(_dbContext);
